refactor: Player domain.player로 패키지 이동
