+++ src/main/java/ladder/domain/line/LadderLinePainter.java
@@ -1,6 +1,6 @@
 package ladder.domain.line;
 
-import ladder.domain.Players;
+import ladder.domain.player.Players;
 import ladder.domain.point.DrawingPointStrategy;
 import ladder.infra.domain.RandomDrawingPointStrategy;
