+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain.player;
 
-import ladder.domain.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
